+++ day12/solution.h
@@ -4,6 +4,6 @@
 // using data_t = char;
 using data_collection_t = charmap_t;
 
-extern long part1(data_collection_t data);
-extern long part2(data_collection_t data);
+extern size_t part1(data_collection_t data);
+extern size_t part2(data_collection_t data);
 extern const data_collection_t read_data(const std::string &path);
